improved Config, fixed adding tracks to database

+++ gui/dialogs/add_files_dialog.py
@@ -18,9 +18,10 @@ class AddFilesDialog(QFileDialog):
         file_paths, _ = self.getOpenFileNames(self,
                                               caption="Add Files to Library",
                                               filter=f"Music Files ({extension_string})")
-        tracks = TracksRepository.convert_file_paths_to_tracks(file_paths)
-        CachedTracksRepository().add_new_tracks(tracks)
-        self.cached_tracks_repository.delete_cache()
-        self.cached_tracks_repository.load_cache()
-        self.added_tracks.emit(tracks)
+        if file_paths:
+            tracks = TracksRepository.convert_file_paths_to_tracks(file_paths)
+            CachedTracksRepository().add_new_tracks(tracks)
+            self.cached_tracks_repository.delete_cache()
+            self.cached_tracks_repository.load_cache()
+            self.added_tracks.emit(tracks)
         self.done(0)
